Extracted some code in cmd/wharf/
